requirements.txt
aiogram==3.4.1
openai-whisper==20231117
ollama==0.1.9
python-dotenv==1.0.1
